bdd-message-fichefrais-fixturesjustificatiffichefrais

+++ src/Entity/FraisForfait.php
@@ -25,7 +25,7 @@ class FraisForfait
     private $libelle;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $montant;
 
@@ -56,12 +56,12 @@ class FraisForfait
         return $this;
     }
 
-    public function getMontant(): ?int
+    public function getMontant(): ?float
     {
         return $this->montant;
     }
 
-    public function setMontant(int $montant): self
+    public function setMontant(float $montant): self
     {
         $this->montant = $montant;
